Change logic for finding gcovr.cfg file
This change modifies the logic when searching for a gcovr.cfg file such
that in the case of an out of source build the current working directory
(i.e. the build directory) will be searched first. If no file is found,
the source root will be searched. This allows OOS build users to
configure their gcovr.cfg file for each build directory if necessary.