Add absolute path to config file filter options
This change modifies the handling of filter options from config files
such that they are always converted to an absolute path. This way, no
matter what the working directory is, the paths to filtered files and
folders will always be interpreted as relative to the config file. This
is handy in out of source builds.